requirements.txt
apache-airflow[postgres]==2.8.1
apache-airflow-providers-postgres==5.11.0
pandas>=2.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
numpy<2.0
alembic>=1.13.0
